requirements/requirements-development.txt
django>=1.8
djangorestframework>=3.1